requirements.txt
Flask==3.1.2
requests==2.32.3
pytest-mock==3.15.1
coverage==7.13.2
pymongo==4.16